feat: add take a break button
Added a 5 minute break button to the pomodoro page. It allows a user to
track the break within the program instead of requiring an external
timer. The break timer is identical to the pomodoro timer except for
duration.

To implement this, refactor `startTimer()` to take a duration in minutes
then add two high-level methods that call `startTimer()` with their
respective durations.