improve: reset timer to duration
When using the short or long break buttons, the pomodoro resets to 25
minutes then changes to 5 or 15 minutes. To the user, the timer goes
from 25:00 to 4:59 or 14:59. This looks odd.

Since `startTimer()` has the duration as a parameter, pass it to
`resetTimer()` so it'll set the clock to the correct duration before
counting down. To keep the existing reset button behavior intact, add a
high-level function that resets with the pomodoro time.